+++ tests/settings.py
@@ -3,8 +3,6 @@ from __future__ import absolute_import, unicode_literals
 
 import os
 
-import django
-
 DEBUG = True
 DEBUG_PROPAGATE_EXCEPTIONS = True
 
@@ -19,11 +17,11 @@ DATABASES = {
     }
 }
 
-INSTALLED_APPS = ["tests.testapp"]
+INSTALLED_APPS = ['django.contrib.contenttypes', "tests.testapp"]
 
-if django.VERSION >= (1, 10):
-    MIDDLEWARE = ()
-else:
-    MIDDLEWARE_CLASSES = ()
+USE_I18N = True
+USE_L10N = True
+
+MIDDLEWARE_CLASSES = ()
 
 SECRET_KEY = "not needed"